WIP: Compute HPWL bottom_up (should be more efficient).

+++ align/schema/constraint.py
@@ -613,7 +613,7 @@ class Boundary(HardConstraint):
 
         if self.max_height is not None:
             assert self.max_height >= 0, f'Boundary:max_height should be greater than zero {self.max_height}'
-            checker.append(checker.cast(bvar.ury-bvar.lly, float) <= 1000*self.max_height, float) # in nanometer
+            checker.append(checker.cast(bvar.ury-bvar.lly, float) <= 1000*self.max_height) # in nanometer
 
 
 class MultiConnection(SoftConstraint):
